+++ worker/include/RTC/SeqManager.hpp
@@ -3,7 +3,7 @@
 
 #include "common.hpp"
 #include <limits> // std::numeric_limits
-#include <set>
+#include <vector>
 
 namespace RTC
 {
@@ -53,7 +53,7 @@ namespace RTC
 		T maxInput{ 0 };
 		T maxDropped{ 0 };
 		T maxForwarded{ 0 };
-		std::set<T, SeqLowerThan> dropped;
+		std::vector<T> dropped;
 	};
 } // namespace RTC