+++ CHANGELOG.md
@@ -2,7 +2,24 @@
 
 ### NEXT
 
+### 3.20.5
+
+- Worker: `SeqManager`, use `std::vector` rather than `std::set` ([PR #1807](https://github.com/versatica/mediasoup/pull/1807)).
+- SCTP: Fixes in `Association.cpp`, `StreamResetHandler.cpp` and `DataTracker.cpp` ([PR #1826](https://github.com/versatica/mediasoup/pull/1826)).
+
+### 3.20.4
+
+- SCTP: Fix `HeartbeatHandler` timers and close the association with `TOO_MANY_RETRIES` error when t3-rtx timer, heartbeat-timeout timer and RE-CONFIG timer expire ([PR #1824](https://github.com/versatica/mediasoup/pull/1824)).
+- SCTP: Fix `ReassemblyQueue::EnterDeferredReset()` ([PR #1825](https://github.com/versatica/mediasoup/pull/1825)).
+
+### 3.20.3
+
+- Fix SCTP crash when the t3-rtx timer expires ([PR #1822](https://github.com/versatica/mediasoup/pull/1822)).
+
+### 3.20.2
+
 - `PortManager`: Replace `uint64_t` hash token with exact-tuple `PortRangeKey` ([PR #1812](https://github.com/versatica/mediasoup/pull/1812), by @999purple999 and @penguinol).
+- Make `DataConsumer.send()` return current buffered amount ([PR #1819](https://github.com/versatica/mediasoup/pull/1819)).
 
 ### 3.20.1